+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+todo
+* use VAR for acronyms list file
+* use database for acronyms
+* add acronym selector in editor
+
 dev - 2023.01.01
 * update to dotclear 2.24
 * use anonymous functions
 * use abstract plugin id
+* use short settings id
 * fix permissions
-* ! dcLegacyEditor and ckEditor are broken !
 
 v1.7.7 - 16-09-2021 Pierre Van Glabeke
 * champs rallongés